Store score in int rather than short to prevent overflow
The score counter overflows regularly with enough rounds.  This is
easily discovered with -P and -s 55440 e.g. to run a battle with
every starting position for warriors written for koenigstuhl-X hill.

+++ pmars-patched/src/global.h
@@ -283,7 +283,7 @@ typedef struct warrior_struct {
   int     instLen;                /* Length of instBank */
   int     offset;                /* Offset value specified by 'ORG' or 'END'.
                                  * 0 is default */
-  short   score[MAXWARRIOR * 2 - 1];
+  int     score[MAXWARRIOR * 2 - 1];
 
   char   *name;                        /* warrior name */
   char   *version;
